+++ src/main/java/com/javampire/openscad/action/ExportAction.java
@@ -27,13 +27,11 @@ public class ExportAction extends OpenSCADExecutableAction {
     @Override
     public void update(@NotNull final AnActionEvent event) {
         super.update(event);
+        final Presentation presentation = event.getPresentation();
+        presentation.setText("Export As ...");
+        presentation.setDescription("Export selected file as ...");
         if (ActionPlaces.EDITOR_TOOLBAR.equals(event.getPlace())) {
-            final Presentation presentation = event.getPresentation();
-            if (presentation.isEnabled()) {
-                presentation.setIcon(AllIcons.Actions.Download);
-                presentation.setText("Export as ...");
-                presentation.setDescription("Export selected file as ...");
-            }
+            presentation.setIcon(AllIcons.Actions.Download);
         }
     }